# docs/swagger.yaml
definitions:
...
  models.ApplicationPolicyRequest:
    properties:
      application_did:
        type: string
      credential:
        additionalProperties: true
        type: object
      issuer_did:
        type: string
      schema_id:
        type: string
    required:
    - application_did
    - credential
    - issuer_did
    - schema_id
    type: object
  models.ApplicationPolicyResponse:
    properties:
      application_did:
        type: string
      credential_id:
        type: string
      issuer_did:
        type: string
      schema_id:
        type: string
    type: object
...
  models.AuthProvider:
    properties:
      app_details:
        $ref: '#/definitions/models.AvailableProvider'
      app_did:
        type: string
      config:
        $ref: '#/definitions/models.OAuthConfig'
    required:
    - app_details
    - app_did
    - config
    type: object
  models.AvailableProvider:
    properties:
      provider_name:
        type: string
      provider_protocol:
        type: string
      provider_type:
        description: social, email, phone, etc
        type: string
    required:
    - provider_name
    - provider_protocol
    - provider_type
    type: object
...
  models.IssueOAuthCredentialRequest:
    properties:
      access_token:
        type: string
      app_did:
        type: string
      provider:
        type: string
      user_did:
        type: string
    required:
    - access_token
    - app_did
    - provider
    - user_did
    type: object
  models.IssueOAuthCredentialResponse:
    properties:
      oauth_credential: {}
      policy_credential: {}
    required:
    - oauth_credential
    - policy_credential
    type: object
  models.JsonSchema:
    properties:
      $schema:
        type: string
      properties:
        additionalProperties: true
        type: object
      required:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
  models.OAuthConfig:
    properties:
      client_id:
        type: string
      redirect_url:
        type: string
    type: object
  models.PolicySchemaRequest:
    properties:
      name:
        type: string
      schema:
        $ref: '#/definitions/models.JsonSchema'
    type: object
  models.PolicySchemaResponse:
    properties:
      id:
        type: string
      name:
        type: string
      schema:
        $ref: '#/definitions/models.JsonSchema'
    type: object
info:
  contact: {}
paths:
...
  /attach-policy:
    post:
      consumes:
      - application/json
      description: Attaches a policy to an application using the provided application
        and issuer DID, and schema ID
      parameters:
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Application Policy Request
        in: body
        name: appPolicy
        required: true
        schema:
          $ref: '#/definitions/models.ApplicationPolicyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully attached policy
          schema:
            $ref: '#/definitions/models.ApplicationPolicyResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Attach policy to application
      tags:
      - Authorization Management
  /auth-provider:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all auth providers
      parameters:
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.AvailableProvider'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get available auth providers
      tags:
      - Authentication Management
  /auth-provider/link:
    post:
      consumes:
      - application/json
      description: Links an OAuth provider to an application by its DID
      parameters:
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Authentication Provider Details
        in: body
        name: provider
        required: true
        schema:
          $ref: '#/definitions/models.AuthProvider'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully linked provider
          schema:
            $ref: '#/definitions/models.AuthProvider'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Link Authentication Provider
      tags:
      - Authentication Management
  /auth-provider/unlink:
    post:
      consumes:
      - application/json
      description: Links an OAuth provider to an application by its DID
      parameters:
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Authentication Provider Details
        in: body
        name: provider
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Successfully linked provider
          schema:
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: UnLink Authentication Provider
      tags:
      - Authentication Management
  /create-policy:
    post:
      consumes:
      - application/json
      description: Creates a new policy based on the provided schema
      parameters:
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      - description: Policy Schema
        in: body
        name: schema
        required: true
        schema:
          $ref: '#/definitions/models.PolicySchemaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully created policy
          schema:
            $ref: '#/definitions/models.PolicySchemaResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create a new policy
      tags:
      - Authorization Management
...
  /issue-credential:
    post:
      consumes:
      - application/json
      description: Issue a new OAuth credential for a user.
      parameters:
      - description: Issue Credential Request
        in: body
        name: issueOAuthCredentialRequest
        required: true
        schema:
          $ref: '#/definitions/models.IssueOAuthCredentialRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Credential successfully issued
          schema:
            $ref: '#/definitions/models.IssueOAuthCredentialResponse'
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Issue OAuth Credential
      tags:
      - Authentication Management
  /policies:
    get:
      consumes:
      - application/json
      description: Retrieves a list of all policies
      parameters:
      - description: API Key
        in: header
        name: x-api-key
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.PolicySchemaResponse'
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get all policies
      tags:
      - Authorization Management
...
  /revoke-credential:
    post:
      consumes:
      - application/json
      description: Revoke an existing OAuth credential.
      parameters:
      - description: Revoke Credential Request
        in: body
        name: revokeOAuthCredentialRequest
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Credential successfully revoked
          schema:
            type: string
        "400":
          description: Bad request
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Revoke OAuth Credential
      tags:
      - Authentication Management
...
swagger: "2.0"